refactor asset pricing service
